calculating next and previous positions

+++ writeopia_ui/src/commonTest/kotlin/io/writeopia/ui/modifiers/StepsModifierTest.kt
@@ -18,10 +18,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.TEXT.type, tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK))),
             StoryStep(type = StoryTypes.TEXT.type, tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK)))
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, 0)
+        val result = StepsModifier.modify(input, 0.0)
             .map { it.storyStep.tags }
             .flatten()
         val expected = listOf(
@@ -47,10 +47,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.TEXT.type, tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK))),
             StoryStep(type = StoryTypes.TEXT.type, tags = setOf(TagInfo(Tag.HIGH_LIGHT_BLOCK)))
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, 0)
+        val result = StepsModifier.modify(input, 0.0)
             .map { it.storyStep.tags }
             .flatten()
             .filter { tagInfo ->
@@ -74,10 +74,10 @@ class StepsModifierTest {
         val input = listOf(
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "code line")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(1, codeBlocks.size)
@@ -91,10 +91,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 1"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 2")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(2, codeBlocks.size)
@@ -109,10 +109,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 2"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 3")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(3, codeBlocks.size)
@@ -128,10 +128,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 2"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 3")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(3, codeBlocks.size)
@@ -147,11 +147,11 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 2"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "line 3")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
         // Use Int.MIN_VALUE to indicate "no drag" - using -1 would cause the first space to be ON_DRAG_SPACE
-        val result = StepsModifier.modify(input, Int.MIN_VALUE)
+        val result = StepsModifier.modify(input, Double.MIN_VALUE)
 
         // Find spaces between code blocks (both SPACE and ON_DRAG_SPACE types)
         val spacesInsideCodeBlock = result.filter {
@@ -186,10 +186,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "block2 line 1"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "block2 line 2")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(4, codeBlocks.size)
@@ -212,10 +212,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "block2 line 1"),
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "block2 line 2")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val codeBlocks = result.filter { it.storyStep.type == StoryTypes.CODE_BLOCK.type }
 
         assertEquals(4, codeBlocks.size)
@@ -235,10 +235,10 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.TEXT.type, text = "text 1"),
             StoryStep(type = StoryTypes.TEXT.type, text = "text 2")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
-        val result = StepsModifier.modify(input, -1)
+        val result = StepsModifier.modify(input, -1.0)
         val types = result.map { it.storyStep.type }
 
         // Text items should have spaces between them
@@ -261,11 +261,11 @@ class StepsModifierTest {
             StoryStep(type = StoryTypes.CODE_BLOCK.type, text = "code 2"),
             StoryStep(type = StoryTypes.TEXT.type, text = "outro")
         ).mapIndexed { index, step ->
-            DrawStory(step, index)
+            DrawStory(step, index.toDouble())
         }
 
         // Use Int.MIN_VALUE to indicate "no drag" - using -1 would cause the first space to be ON_DRAG_SPACE
-        val result = StepsModifier.modify(input, Int.MIN_VALUE)
+        val result = StepsModifier.modify(input, Double.MIN_VALUE)
 
         // Get non-space items
         val contentItems = result.filter {
